[Refactor] Resume 페이지 컴포넌트화 및 일부 구조 변경
