+++ src/pages/Resume/constants/etc.js
@@ -1,0 +1,22 @@
+export const EDUCATION = [
+  {
+    period: '2012.03 ~ 2018.08',
+    title: '영남대학교',
+    explanation:
+      '환경공학 전공 / 산업안전기사, 대기환경기사 ,수질환경기사 취득',
+  },
+  {
+    period: '2009.03 ~ 2012.02',
+    title: '달성고등학교',
+    explanation: '자연계 졸업 (대구광역시)',
+  },
+];
+
+export const ETC = [
+  {
+    period: '2018.10 ~ 2020.05',
+    title: '베올리아 코리아',
+    explanation:
+      '수처리(폐수 초순수) 엔지니어 근무 후 사업개발 팀 / SK하이닉스 계약 담당으로 기술 기획 업무 진행',
+  },
+];
