+++ src/pages/Resume/constants/links.js
@@ -1,0 +1,22 @@
+export const PROFILE_LINK = [
+  {
+    icon: 'fab fa-github',
+    href: 'https://github.com/sanghunlee-711',
+    name: 'https://github.com/sanghunlee-711',
+  },
+  {
+    icon: 'fab fa-google',
+    href: 'mailto:cloudlee711@gmail.com',
+    name: 'cloudlee711@gmail.com',
+  },
+  {
+    icon: 'fas fa-blog',
+    href: 'https://blog.cloud-sanghun.com',
+    name: 'https://blog.cloud-sanghun.com',
+  },
+  {
+    icon: 'fas fa-briefcase',
+    href: 'https://blog.cloud-sanghun.com/portfolio',
+    name: 'Portfolio',
+  },
+];
