+++ src/pages/Resume/constants/skills.js
@@ -1,0 +1,50 @@
+export const WELL_KNOWN_LIST = [
+  'ES5/ES6',
+  'TypeScript',
+  'React.js',
+  'Next.js',
+  'Redux',
+  'Redux-Toolkit',
+  'Redux-saga',
+  'styled-components',
+];
+
+export const SKILL = [
+  {
+    title: 'CORE',
+    skill: ['ES5/ES6', 'TypeScript', 'React.js', 'Next.js'],
+  },
+  {
+    title: 'STATE MANAGEMENT',
+    skill: [
+      'Redux',
+      'Redux-Toolkit',
+      'Redux-saga',
+      'Apollo-client',
+      'GraphQL',
+      'React-Query',
+    ],
+  },
+  {
+    title: 'BUILD SYSTEM',
+    skill: ['Webpack', 'Rollup.js'],
+  },
+  {
+    title: 'STYLE',
+    skill: ['styled-components', 'SCSS'],
+  },
+  {
+    title: 'TEST',
+    skill: ['Cypress', 'Jest', 'testing-library', 'msw'],
+  },
+  { title: 'SERVER', skill: ['Express.js', 'Nest.js', 'Nginx', 'type-orm'] },
+  { title: 'DATABASE', skill: ['Postgresql', 'MySQL'] },
+  {
+    title: 'CI/CD',
+    skill: ['Jenkins'],
+  },
+  {
+    title: 'ETC',
+    skill: ['Git', 'Jira'],
+  },
+];
